+++ packages/web3-core/src/web3_subscription_manager.ts
@@ -39,6 +39,17 @@ export class Web3SubscriptionManager<
 		InstanceType<RegisteredSubs[keyof RegisteredSubs]>
 	> = new Map();
 
+	/**
+	 *
+	 * @param requestManager
+	 * @param registeredSubscriptions
+	 *
+	 * @example
+	 * ```ts
+	 * const requestManager = new Web3RequestManager("ws://localhost:8545");
+	 * const subscriptionManager = new Web3SubscriptionManager(requestManager, {});
+	 * ```
+	 */
 	public constructor(
 		public readonly requestManager: Web3RequestManager<API>,
 		public readonly registeredSubscriptions: RegisteredSubs,
@@ -52,6 +63,16 @@ export class Web3SubscriptionManager<
 		});
 	}
 
+	/**
+	 * Will create a new subscription
+	 *
+	 * @param name - The subscription you want to subscribe to
+	 * @param args (optional) - Optional additional parameters, depending on the subscription type
+	 * @param returnFormat ({@link DataFormat} defaults to {@link DEFAULT_RETURN_FORMAT}) - Specifies how the return data from the call should be formatted.
+	 *
+	 * Will subscribe to a specific topic (note: name)
+	 * @returns The subscription object
+	 */
 	public async subscribe<T extends keyof RegisteredSubs>(
 		name: T,
 		args?: ConstructorParameters<RegisteredSubs[T]>[0],
@@ -76,10 +97,19 @@ export class Web3SubscriptionManager<
 		return subscription;
 	}
 
+	/**
+	 * Will returns all subscriptions.
+	 */
 	public get subscriptions() {
 		return this._subscriptions;
 	}
 
+	/**
+	 *
+	 * Adds an instance of {@link Web3Subscription} and subscribes to it
+	 *
+	 * @param sub - A {@link Web3Subscription} object
+	 */
 	public async addSubscription(sub: InstanceType<RegisteredSubs[keyof RegisteredSubs]>) {
 		if (!this.supportsSubscriptions()) {
 			throw new SubscriptionError('The current provider does not support subscriptions');
@@ -97,6 +127,11 @@ export class Web3SubscriptionManager<
 
 		this._subscriptions.set(sub.id, sub);
 	}
+	/**
+	 * Will clear a subscription
+	 *
+	 * @param id - The subscription of type {@link Web3Subscription}  to remove
+	 */
 
 	public async removeSubscription(sub: InstanceType<RegisteredSubs[keyof RegisteredSubs]>) {
 		if (isNullish(sub.id)) {
@@ -115,7 +150,12 @@ export class Web3SubscriptionManager<
 		this._subscriptions.delete(id);
 		return id;
 	}
-
+	/**
+	 * Will unsubscribe all subscriptions that fulfill the condition
+	 *
+	 * @param condition - A function that access and `id` and a `subscription` and return `true` or `false`
+	 * @returns An array of all the un-subscribed subscriptions
+	 */
 	public async unsubscribe(condition?: ShouldUnsubscribeCondition) {
 		const result = [];
 		for (const [id, sub] of this.subscriptions.entries()) {
@@ -127,10 +167,18 @@ export class Web3SubscriptionManager<
 		return Promise.all(result);
 	}
 
+	/**
+	 * Clears all subscriptions
+	 */
 	public clear() {
 		this._subscriptions.clear();
 	}
 
+	/**
+	 * Check whether the current provider supports subscriptions.
+	 *
+	 * @returns `true` or `false` depending on if the current provider supports subscriptions
+	 */
 	public supportsSubscriptions(): boolean {
 		return isNullish(this.requestManager.provider)
 			? false