+++ packages/web3-eth/test/integration/block/rpc.getUncle.test.ts
@@ -41,6 +41,8 @@ describe('rpc with block', () => {
 		earliest: 'earliest';
 		latest: 'latest';
 		pending: 'pending';
+		finalized: 'finalized';
+		safe: 'safe';
 		blockNumber: number | bigint;
 		blockHash: string;
 		transactionHash: string;
@@ -84,6 +86,8 @@ describe('rpc with block', () => {
 			pending: 'pending',
 			latest: 'latest',
 			earliest: 'earliest',
+			finalized: 'finalized',
+			safe: 'safe',
 			blockNumber: Number(receipt.blockNumber),
 			blockHash: String(receipt.blockHash),
 			transactionHash: String(receipt.transactionHash),
@@ -98,9 +102,9 @@ describe('rpc with block', () => {
 	describe('methods', () => {
 		it.each(
 			toAllVariants<{
-				block: 'earliest' | 'latest' | 'pending' | 'blockNumber';
+				block: 'earliest' | 'latest' | 'pending' | 'finalized' | 'safe' | 'blockNumber';
 			}>({
-				block: ['earliest', 'latest', 'pending', 'blockNumber'],
+				block: ['earliest', 'latest', 'pending', 'finalized', 'safe', 'blockNumber'],
 			}),
 		)('getUncle', async ({ block }) => {
 			const res = await web3Eth.getUncle(blockData[block], 0);