+++ packages/web3-eth-contract/CHANGELOG.md
@@ -228,3 +228,4 @@ const transactionHash = receipt.transactionHash;
 
 -   Added functionality of `createAccessList` for contracts ( #5780 )
 -   An instance of `Contract` will `subscribeToContextEvents` upon instantiation if `syncWithContext` is set to `true` and the constructor is passed an instance of `Web3Context` (#5833)
+-   Added support of `safe` and `finalized` block tags (#5823)